# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interview-result/{sessionId}:
    get:
      operationId: getInterviewResult
      tags: [interview]
      summary: Retrieve a completed interview result by session ID
      description: |
        Returns the completed interview record stored in the database for the given session ID.
        For authenticated-user sessions, the caller must be the same user who conducted the interview.
        For guest sessions, the caller must hold the per-session claim cookie set when the session was created.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed interview record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewResultRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewError"
...
  /interview-sessions:
    get:
      operationId: listInterviewSessions
      tags: [sessions]
      summary: List all interview sessions for the authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of stored interview sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewSessionList'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewError'
  /interview-sessions/{sessionId}:
    put:
      operationId: saveInterviewSession
      tags: [sessions]
      summary: Save or update an interview session
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewSessionData'
      responses:
        '200':
          description: Session saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewSessionData'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewError'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    InterviewMessage:
      type: object
      properties:
        role:
          type: string
          enum: [student, advisor]
        content:
          type: string
      required:
        - role
        - content
...
    InterviewProgress:
      type: object
      description: Subtle, non-checklist progress signal (0-100).
      properties:
        percent:
          type: integer
          minimum: 0
          maximum: 100
        stage:
          type: string
          description: A short human-readable stage label.
      required:
        - percent
        - stage
    InterviewRecommendation:
      type: object
      properties:
        recommendedMajor:
          type: string
        matchScore:
          type: integer
          minimum: 0
          maximum: 100
        whyItFits:
          type: array
          items:
            type: string
        alternativeMajors:
          type: array
          items:
            type: string
        academicStrengths:
          type: array
          items:
            type: string
        careerAdvice:
          type: array
          items:
            type: string
        closingMessage:
          type: string
        admissionNote:
          type: ["string", "null"]
          description: Optional note about AAUP admission requirements and GPA eligibility.
      required:
        - recommendedMajor
        - matchScore
        - whyItFits
        - alternativeMajors
        - academicStrengths
        - careerAdvice
        - closingMessage
...
    InterviewError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    InterviewResultRecord:
      type: object
      description: A completed interview record stored in the database.
      properties:
        sessionId:
          type: ["string", "null"]
          description: The session ID this record is linked to.
        recommendation:
          $ref: "#/components/schemas/InterviewRecommendation"
        fullConversation:
          type: array
          items:
            $ref: "#/components/schemas/InterviewMessage"
        savedAt:
          type: string
          format: date-time
        user:
          type: ["object", "null"]
          properties:
            id:
              type: string
            email:
              type: ["string", "null"]
            firstName:
              type: ["string", "null"]
            lastName:
              type: ["string", "null"]
    InterviewSessionData:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        title:
          type: ["string", "null"]
        profileData:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional profile form data captured at each step.
        messages:
          type: array
          items:
            $ref: "#/components/schemas/InterviewMessage"
        progress:
          oneOf:
            - $ref: "#/components/schemas/InterviewProgress"
            - type: "null"
        recommendation:
          oneOf:
            - $ref: "#/components/schemas/InterviewRecommendation"
            - type: "null"
      required:
        - id
        - createdAt
        - updatedAt
        - messages
    InterviewSessionList:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/InterviewSessionData"
      required:
        - sessions
...