# lib/api-zod/src/generated/types/interviewResultRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewMessage } from "./interviewMessage";
import type { InterviewRecommendation } from "./interviewRecommendation";
import type { InterviewResultRecordUser } from "./interviewResultRecordUser";

/**
 * A completed interview record stored in the database.
 */
export interface InterviewResultRecord {
  /**
   * The session ID this record is linked to.
   * @nullable
   */
  sessionId?: string | null;
  recommendation?: InterviewRecommendation;
  fullConversation?: InterviewMessage[];
  savedAt?: Date;
  /** @nullable */
  user?: InterviewResultRecordUser;
}