# lib/api-zod/src/generated/types/interviewSessionData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewMessage } from "./interviewMessage";
import type { InterviewProgress } from "./interviewProgress";
import type { InterviewRecommendation } from "./interviewRecommendation";
import type { InterviewSessionDataProfileData } from "./interviewSessionDataProfileData";

export interface InterviewSessionData {
  id: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  title?: string | null;
  /** Optional profile form data captured at each step. */
  profileData?: InterviewSessionDataProfileData;
  messages: InterviewMessage[];
  progress?: InterviewProgress | null;
  recommendation?: InterviewRecommendation | null;
}